+++ simulations_study/multivariate_data_generation_two_bases.R
@@ -173,7 +173,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["128"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["128"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 128) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["128"]], 
                          params = list(shifts = list(edges = shifts_grid[["128_3"]]$edges, values = shifts_grid[["128_3"]]$values[1, ])),
                          adj.root = 0,
@@ -203,7 +203,7 @@ extract.parsimonyNumber(parsimonyNumber(trees[["128"]], clusters_from_shifts(tre
 # vec_Y <- kronecker(T_tree[["128"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 # X1.tips.exp.mat <- matrix(vec_Y, p_base, ntaxa_base) + beta_0
 # unique(X1.tips.exp.mat[1, ])
-# plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+# plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
 #                          phylo = trees[["128"]], 
 #                          params = list(shifts = list(edges = shifts_grid[["128_7"]]$edges, values = shifts_grid[["128_7"]]$values[1, ])),
 #                          adj.root = 0,
@@ -238,7 +238,7 @@ extract.parsimonyNumber(parsimonyNumber(trees[["128"]], clusters_from_shifts(tre
 # vec_Y <- kronecker(T_tree[["128"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 # X1.tips.exp.mat <- matrix(vec_Y, p_base, ntaxa_base) + beta_0
 # unique(X1.tips.exp.mat[1, ])
-# plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+# plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
 #                          phylo = trees[["128"]], 
 #                          params = list(shifts = list(edges = shifts_grid[["128_11"]]$edges, values = shifts_grid[["128_11"]]$values[1, ])),
 #                          adj.root = 0,
@@ -278,7 +278,7 @@ extract.parsimonyNumber(parsimonyNumber(trees[["128"]], clusters_from_shifts(tre
 # vec_Y <- kronecker(T_tree[["128"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 # X1.tips.exp.mat <- matrix(vec_Y, p_base, ntaxa_base) + beta_0
 # unique(X1.tips.exp.mat[1, ])
-# plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+# plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
 #                          phylo = trees[["128"]], 
 #                          params = list(shifts = list(edges = shifts_grid[["128_15"]]$edges, values = shifts_grid[["128_15"]]$values[1, ])),
 #                          adj.root = 0,
@@ -304,7 +304,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["32"]], alpha_base * diag(1
 vec_Y <- kronecker(T_tree[["32"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 32) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["32"]], 
                          params = list(shifts = list(edges = shifts_grid[["32_3"]]$edges, values = shifts_grid[["32_3"]]$values[1, ])),
                          adj.root = 0,
@@ -330,7 +330,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["64"]], alpha_base * diag(1
 vec_Y <- kronecker(T_tree[["64"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 64) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["64"]], 
                          params = list(shifts = list(edges = shifts_grid[["64_3"]]$edges, values = shifts_grid[["64_3"]]$values[1, ])),
                          adj.root = 0,
@@ -356,7 +356,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["96"]], alpha_base * diag(1
 vec_Y <- kronecker(T_tree[["96"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 96) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["96"]], 
                          params = list(shifts = list(edges = shifts_grid[["96_3"]]$edges, values = shifts_grid[["96_3"]]$values[1, ])),
                          adj.root = 0,
@@ -382,7 +382,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["160"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["160"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 160) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["160"]], 
                          params = list(shifts = list(edges = shifts_grid[["160_3"]]$edges, values = shifts_grid[["160_3"]]$values[1, ])),
                          adj.root = 0,
@@ -412,7 +412,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["160"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["160"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 160) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["160"]], 
                          params = list(shifts = list(edges = shifts_grid[["160_7"]]$edges, values = shifts_grid[["160_7"]]$values[1, ])),
                          adj.root = 0,
@@ -447,7 +447,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["160"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["160"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 160) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["160"]], 
                          params = list(shifts = list(edges = shifts_grid[["160_11"]]$edges, values = shifts_grid[["160_11"]]$values[1, ])),
                          adj.root = 0,
@@ -487,7 +487,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["160"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["160"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 160) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["160"]], 
                          params = list(shifts = list(edges = shifts_grid[["160_15"]]$edges, values = shifts_grid[["160_15"]]$values[1, ])),
                          adj.root = 0,
@@ -513,7 +513,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["192"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["192"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 192) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["192"]], 
                          params = list(shifts = list(edges = shifts_grid[["192_3"]]$edges, values = shifts_grid[["192_3"]]$values[1, ])),
                          adj.root = 0,
@@ -539,7 +539,7 @@ W <- compute_actualization_matrix_ultrametric(trees[["256"]], alpha_base * diag(
 vec_Y <- kronecker(T_tree[["256"]], diag(1, p_base, p_base)) %*% W %*% as.vector(Delta)
 X1.tips.exp.mat <- matrix(vec_Y, p_base, 256) + beta_0
 unique(X1.tips.exp.mat[1, ])
-plot.data.process.actual(Y.state = X1.tips.exp.mat[1, ],
+plot_data.process.actual(Y.state = X1.tips.exp.mat[1, ],
                          phylo = trees[["256"]], 
                          params = list(shifts = list(edges = shifts_grid[["256_3"]]$edges, values = shifts_grid[["256_3"]]$values[1, ])),
                          adj.root = 0,