+++ NEWS.md
@@ -1,6 +1,13 @@
-# PhylogeneticEM 1.6.0.9000
+# PhylogeneticEM 1.7.0.9000
 * Bug fix:
   * `merge_rotation` now works for trait with more than 2 dimensions.
+
+# PhylogeneticEM 1.7.0
+* Bug fixes:
+  * Fix documentation notes on CRAN checks.
+* New functions and features:
+  * Function `node_optimal_values` for optimal values computation at each nodes.
+  * Trait names are now inherited by the `params_process` objects.
 
 # PhylogeneticEM 1.6.0
 * Deprecation fix: